backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# AI and Machine Learning
openai==1.6.1

# LangChain dependencies (may not install on Python 3.13)
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector database and embeddings
faiss-cpu==1.7.4
chromadb==0.4.18

# NLP and text processing
tiktoken==0.5.2
Levenshtein==0.25.0

# ML dependencies
scikit-learn==1.3.2
sentence-transformers==2.2.2
numpy==1.24.3

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Resume processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.9.0

# PDF generation
reportlab==4.0.7

# Security Dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
redis==4.5.4
python-magic==0.4.27
psutil==5.9.6
pydantic-settings>=2.4.0

# Email Service
fastapi-mail==1.4.1
jinja2==3.1.2

# Social Authentication
httpx==0.27.0
python-jose[cryptography]==3.3.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Additional Security & Monitoring
sentry-sdk[fastapi]==1.38.0
cryptography==41.0.7

# Authentication & Database (added for user auth)
email-validator==2.1.0
SQLAlchemy==2.0.36
alembic==1.14.0

# Payment Processing
stripe==8.0.0

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0